fix edit-view and add simple_form
