+++ spovm_2/lab1/lib/IP.rb
@@ -1,0 +1,20 @@
+require 'bit-struct'
+
+class IP < BitStruct
+  unsigned    :ip_v,     4,     "Version"
+  unsigned    :ip_hl,    4,     "Header length"
+  unsigned    :ip_tos,   8,     "TOS"
+  unsigned    :ip_len,  16,     "Length"
+  unsigned    :ip_id,   16,     "ID"
+  unsigned    :ip_off,  16,     "Frag offset"
+  unsigned    :ip_ttl,   8,     "TTL"
+  unsigned    :ip_p,     8,     "Protocol"
+  unsigned    :ip_sum,  16,     "Checksum"
+  octets      :ip_src,  32,     "Source addr"
+  octets      :ip_dst,  32,     "Dest addr"
+  rest        :body,            "Body of message"
+
+  note "     rest is application defined message body"
+  initial_value.ip_v    = 4
+  initial_value.ip_hl   = 5
+end
